# docs/swagger.yaml
definitions:
  models.Article:
    description: Article db model
    properties:
      body:
        type: string
      id:
        type: integer
      recordID:
        type: integer
    type: object
  models.ArticleCreate:
    description: Article create model
    properties:
      body:
        example: '{html page}'
        format: html
        type: string
    required:
    - body
    type: object
  models.ArticleDto:
    description: Article dto model
    properties:
      body:
        example: '{html page}'
        format: html
        type: string
      id:
        example: 1
        type: integer
    required:
    - body
    - id
    type: object
  models.Record:
    properties:
      article:
        $ref: '#/definitions/models.Article'
      hidden:
        type: boolean
      id:
        type: integer
      status:
        $ref: '#/definitions/models.RecordStatus'
      title:
        type: string
      userID:
        type: integer
      videoLink:
        type: string
    type: object
  models.RecordCreate:
    description: Record create model
    properties:
      title:
        example: title
        type: string
      video_link:
        example: https://www.youtube.com/watch?v=4O3UGW-Bbbw
        format: url
        type: string
    required:
    - title
    - video_link
    type: object
  models.RecordDto:
    description: Record dto model
    properties:
      hidden:
        example: true
        type: boolean
      id:
        example: 1
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/models.RecordStatus'
        example: В обработке
      title:
        example: title
        type: string
      video_link:
        example: https://www.youtube.com/watch?v=4O3UGW-Bbbw
        format: url
        type: string
    required:
    - hidden
    - id
    - status
    - title
    - video_link
    type: object
  models.RecordStatus:
    enum:
    - processin
    - completed
    - published
    type: string
    x-enum-varnames:
    - ProcessingRecordStatus
    - CompletedRecordStatus
    - PublishedRecordStatus
  models.TokenResponse:
    description: Token response model
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        type: string
      token_type:
        example: Bearer
        type: string
    required:
    - token
    - token_type
    type: object
  models.User:
    description: User db model
    properties:
      email:
        type: string
      fio:
        type: string
      id:
        type: integer
      password:
        type: string
      records:
        items:
          $ref: '#/definitions/models.Record'
        type: array
    type: object
  models.UserCreate:
    description: User create model
    properties:
      email:
        example: test@test.ru
        format: email
        type: string
      fio:
        example: Мисосов Мисос Мисосович
        type: string
      password:
        example: test
        type: string
    required:
    - email
    - fio
    - password
    type: object
  models.UserLogin:
    description: User login model
    properties:
      email:
        example: test@test.ru
        format: email
        type: string
      password:
        example: test
        type: string
    required:
    - email
    - password
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is a sample server for Profbuh API.
  title: Profbuh API
paths:
  /api/v1/test:
    get:
      consumes:
      - application/json
      description: Test middleware
      produces:
      - application/json
      responses:
        "200":
          description: Current user
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - Bearer: []
      summary: Test middleware
      tags:
      - test
  /article/{record_id}:
    get:
      consumes:
      - application/json
      description: Get article for record by record_id
      parameters:
      - description: Record id
        in: path
        name: record_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Article
          schema:
            $ref: '#/definitions/models.ArticleDto'
        "400":
          description: Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get article
      tags:
      - article
  /article/create:
    post:
      consumes:
      - application/json
      description: Create article
      parameters:
      - description: Record id
        in: query
        name: record_id
        required: true
        type: integer
      - description: Article create info
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/models.ArticleCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ArticleDto'
        "400":
          description: Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create article
      tags:
      - article
  /auth/user/create:
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: User create info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created token for user
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Create user
      tags:
      - user
  /auth/user/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: User login info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            $ref: '#/definitions/models.TokenResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Login user
      tags:
      - user
  /record/{record_id}:
    get:
      consumes:
      - application/json
      description: Get record by id
      parameters:
      - description: Record id
        in: path
        name: record_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Record
          schema:
            $ref: '#/definitions/models.RecordDto'
        "400":
          description: Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get record by id
      tags:
      - record
  /record/{record_id}/publish:
    post:
      consumes:
      - application/json
      description: Publish record
      parameters:
      - description: Record id
        in: path
        name: record_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RecordDto'
        "400":
          description: Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Publish record
      tags:
      - record
  /record/all:
    get:
      consumes:
      - application/json
      description: Get all records(not articles) for current user
      produces:
      - application/json
      responses:
        "200":
          description: User's records
          schema:
            items:
              $ref: '#/definitions/models.RecordDto'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get records by user
      tags:
      - record
  /record/create:
    post:
      consumes:
      - application/json
      description: Create record
      parameters:
      - description: Record create info
        in: body
        name: record
        required: true
        schema:
          $ref: '#/definitions/models.RecordCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Created record
          schema:
            $ref: '#/definitions/models.RecordDto'
        "400":
          description: Bad request
          schema:
            type: string
      security:
      - Bearer: []
      summary: Create record
      tags:
      - record
swagger: "2.0"